[commented out] transporter.sendMail for nodemailer in app.js for now to prevent emails from sending whenever changes are made
